Extract messageTimeout to state module
