+++ consensus/ibft/state_test.go
@@ -2,6 +2,7 @@ package ibft
 
 import (
 	"testing"
+	"time"
 
 	"github.com/dogechain-lab/dogechain/consensus/ibft/proto"
 	"github.com/dogechain-lab/dogechain/types"
@@ -126,6 +127,101 @@ func TestState_PorposerAndNeedPunished(t *testing.T) {
 
 			punished := state.CalcNeedPunished(tt.round, tt.lastBlockProposer)
 			assert.Equal(t, tt.needPunished, punished)
+		})
+	}
+}
+
+func TestState_MessageTimeout(t *testing.T) {
+	// fake addr
+	var addr1 = types.StringToAddress("1")
+
+	//nolint:lll
+	testCases := []struct {
+		description string
+		c           *currentState
+		expected    time.Duration
+	}{
+		{
+			description: "for 0 validator returns 4s",
+			c: &currentState{
+				validators: ValidatorSet{},
+			},
+			expected: 4 * time.Second,
+		},
+		{
+			description: "for 1 validator returns 4s",
+			c: &currentState{
+				validators: ValidatorSet{
+					addr1,
+				}},
+			expected: 4 * time.Second,
+		},
+		{
+			description: "for 2 validators returns 4s",
+			c: &currentState{
+				validators: ValidatorSet{
+					addr1, addr1,
+				}},
+			expected: 4 * time.Second,
+		},
+		{
+			description: "for 3 validators returns 6s",
+			c: &currentState{
+				validators: ValidatorSet{
+					addr1, addr1, addr1,
+				}},
+			expected: 6 * time.Second,
+		},
+		{
+			description: "for 13 validators returns 12s",
+			c: &currentState{
+				validators: ValidatorSet{
+					addr1, addr1, addr1, addr1, addr1, addr1, addr1,
+					addr1, addr1, addr1, addr1, addr1, addr1,
+				}},
+			expected: 12 * time.Second,
+		},
+		{
+			description: "for 23 validators returns 18s",
+			c: &currentState{
+				validators: ValidatorSet{
+					addr1, addr1, addr1, addr1, addr1, addr1, addr1,
+					addr1, addr1, addr1, addr1, addr1, addr1, addr1,
+					addr1, addr1, addr1, addr1, addr1, addr1, addr1,
+					addr1, addr1,
+				}},
+			expected: 18 * time.Second,
+		},
+		{
+			description: "for 24 validators returns 20s",
+			c: &currentState{
+				validators: ValidatorSet{
+					addr1, addr1, addr1, addr1, addr1, addr1, addr1,
+					addr1, addr1, addr1, addr1, addr1, addr1, addr1,
+					addr1, addr1, addr1, addr1, addr1, addr1, addr1,
+					addr1, addr1, addr1,
+				}},
+			expected: 20 * time.Second,
+		},
+		{
+			description: "for 28 validators returns 20s",
+			c: &currentState{
+				validators: ValidatorSet{
+					addr1, addr1, addr1, addr1, addr1, addr1, addr1,
+					addr1, addr1, addr1, addr1, addr1, addr1, addr1,
+					addr1, addr1, addr1, addr1, addr1, addr1, addr1,
+					addr1, addr1, addr1, addr1, addr1, addr1, addr1,
+				}},
+			expected: 20 * time.Second,
+		},
+	}
+
+	for _, test := range testCases {
+		test := test
+		t.Run(test.description, func(t *testing.T) {
+			timeout := test.c.messageTimeout()
+
+			assert.Equal(t, test.expected, timeout)
 		})
 	}
 }
